Added individual config class for opentracing domain
